geo-rep: Re Enable ignore_deletes Option
If this option is set, Deletes will not be propogated to Slave.
This option is applicable for UNLINK and RMDIR.

gluster volume geo-replication <MASTER> <SLAVEHOST>::<SLAVEVOL> \
        config ignore_deletes true

Default value is false.

PS: Use this option with caution, If you create the file in master
with same path then it fails to sync to slave. Old file in Slave
will have different GFID compared to New.

BUG: 1189363
Reviewed-on: http://review.gluster.org/9583
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ geo-replication/syncdaemon/master.py
@@ -946,8 +946,9 @@ class GMasterChangelogMixin(GMasterCommon):
                     if pt in datas:
                         datas.remove(pt)
 
-                    purge_update()
-                    entries.append(edct(ty, gfid=gfid, entry=en))
+                    if not boolify(gconf.ignore_deletes):
+                        purge_update()
+                        entries.append(edct(ty, gfid=gfid, entry=en))
                 elif ty in ['CREATE', 'MKDIR', 'MKNOD']:
                     entry_update()
                     # stat information present in the changelog itself
